refactor: `AuthController` 패키지 이동
